Cleaned up, fixed broken shading on -vs axis
removed references to binormal and tangent vectors in the VertexType when
sending to the GPU. Since this is only needed for the plane they are hard
coded into the vertex shader.

replaced the floor normal textures as it appeared to have broken values
which was at least part of the cause of the lighting bug.

There are still a couple of TODOs sprinkled around and commented out code
that could be removed for maximum cleanlines. Also the whole binormal
tangent thing could be calculated somewher else, included in the model,
maybe generated in the geometry shader. Anywhere really would be better
than having it hardcoded into the vertex shader but it works for now.